+++ main.go
@@ -58,11 +58,13 @@ func main() {
 
 func initRestyClient() {
 	extalertrules.RestyClient = resty.New()
+	extalertrules.RestyClient.SetTimeout(config.Config.GetApiTimeout())
 	extalertrules.RestyClient.SetBaseURL(config.Config.ApiBaseUrl)
 	extalertrules.RestyClient.SetHeader("Authorization", "Bearer "+config.Config.ServiceToken)
 	extalertrules.RestyClient.SetHeader("Content-Type", "application/json")
 
 	extannotations.RestyClient = resty.New()
+	extannotations.RestyClient.SetTimeout(config.Config.GetApiTimeout())
 	extannotations.RestyClient.SetRetryCount(2)
 	extannotations.RestyClient.SetRetryWaitTime(500 * time.Millisecond)
 	extannotations.RestyClient.SetBaseURL(config.Config.ApiBaseUrl)